+++ app/Http/Controllers/Backend/PostController.php
@@ -14,9 +14,12 @@ use Illuminate\Support\Facades\File;
 use Illuminate\Support\Facades\Session;
 
 class PostController extends Controller{
-	public function index(){
+	public function index(Request $request){
+		$this->authorize('viewAny',Post::class);
 		$posts = Post::with('news_category:id,name')->withTrashed();
-		$posts = $posts->orderBy('updated_at','DESC')->paginate(4);
+		if($request->title) $posts->where('title','like','%'.$request->title.'%');
+		if($request->news_category_id) $posts->where('news_category_id',$request->news_category_id);
+		$posts = $posts->orderBy('updated_at','DESC')->paginate(8);
 		$news_categories = NewsCategory::all();
 		return view('backend.posts.index')->with([
 			'posts' => $posts,
@@ -25,6 +28,7 @@ class PostController extends Controller{
 	}
 
     public function create(){
+    	$this->authorize('create',Post::class);
     	$news_categories = NewsCategory::all();
     	return view('backend.posts.create')->with([
     		'news_categories' => $news_categories,
@@ -32,6 +36,7 @@ class PostController extends Controller{
     }
 
     public function store(RequestPost $requestPost){
+    	$this->authorize('create',Post::class);
     	$post = new Post();
 		$user = Auth::user();
 
@@ -65,6 +70,7 @@ class PostController extends Controller{
 
 	public function edit($id){
 		$post = Post::withTrashed()->findOrFail($id);
+		$this->authorize('update',$post);
 		$news_categories = NewsCategory::all();
 		return view('backend.posts.edit')->with([
 			'post' => $post,
@@ -74,7 +80,7 @@ class PostController extends Controller{
 
 	public function update(RequestPost $requestPost,$id){
 		$post = Post::withTrashed()->findOrFail($id);
-
+		$this->authorize('update',$post);
 		$post->title = $requestPost->get('title');
 		$post->description = $requestPost->get('description');
 		$post->content = $requestPost->get('content');
@@ -106,12 +112,14 @@ class PostController extends Controller{
 
 	public function destroy($id){
 		$post = Post::findOrFail($id);
+		$this->authorize('delete',$post);
 		$post->delete();
 		return redirect()->route('backend.post.index');
 	}
 
 	public function forceDelete($id){
 		$post = Post::onlyTrashed()->findOrFail($id);
+		$this->authorize('forceDelete',$post);
 		File::delete($post->image);
 		File::delete($post->background_img_title);
 		$post->forceDelete();
@@ -120,13 +128,14 @@ class PostController extends Controller{
 
 	public function restore($id){
 		$post = Post::onlyTrashed()->findOrFail($id);
+		$this->authorize('restore',$post);
 		$post->restore();
 		return redirect()->route('backend.post.index');
 	}
 
 	public function changeHot($id){
 		$post = Post::withTrashed()->findOrFail($id);
-
+		$this->authorize('changeHot',$post);
 		if(Auth::user()->can('changeHot',$post)){
 	    	$post->hot = !$post->hot;
 	    	$post->update();
